Change loop_rate to support fractional values
Previously, loop_rate was limited to integer values, restricting the Rate
object to whole-second intervals. By changing the data type to double, we
enable fractional  rates, allowing the Rate object to sleep for
intervals longer than 1 second or for precise fractional durations.

+++ src/iio_attr_topic.cpp
@@ -20,7 +20,7 @@
 
 IIOAttrTopic::IIOAttrTopic(
   std::shared_ptr<IIONode> nh, std::string topicName, std::string attrPath,
-  topicType_t type, int loopRate)
+  topicType_t type, double loopRate)
 : m_loopRate(loopRate)
 {
   RCLCPP_INFO(rclcpp::get_logger("adi_iio_node"), "Created IIOAttrTopic");
